refactor: 쿠키에서 토큰을 추출하는 로직 JwtTokenProvider로 위치 변경
